requirements.txt
# Core LangChain and LLM dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
openai==1.7.2
anthropic==0.8.1

# Additional AI/ML libraries
tiktoken==0.5.2
pydantic==2.5.3
pydantic-settings==2.1.0

# Resume parsing and document processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
spacy==3.7.2

# API and web framework
fastapi==0.108.0
uvicorn==0.25.0
requests==2.31.0
httpx==0.26.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Logging and monitoring
structlog==24.1.0
